% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distances.R
\name{coordinate_grids}
\alias{coordinate_grids}
\title{Creates RasterLayer objects filled by the x and y grid coordinates}
\usage{
coordinate_grids(object)
}
\arguments{
\item{object}{RasterLayer object}
}
\value{
RasterStack object
}
\description{
Creates RasterLayer objects filled by the x and y grid coordinates
}
