% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{cross_validate}
\alias{cross_validate}
\title{K-Fold cross-validation scheme}
\usage{
cross_validate(object, formula, data, param_grid = NULL, n_splits = 10,
  keep_models = FALSE)
}
\arguments{
\item{object}{partial function of ranger}

\item{formula}{model forumula}

\item{data}{data.frame of training data}

\item{param_grid}{pass a function that generates an expand.grid of
hyperparameters}

\item{n_splits}{integer, number of cross validation folds}

\item{keep_models}{logical, save each model to the tibble per cross
validation fold. Note that this can consume a lot of memory for large
datasets}
}
\value{
tibble containing cross validation results
}
\description{
K-Fold cross-validation scheme
}
