% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distances.R
\name{dist_to_categories}
\alias{dist_to_categories}
\title{Calculate grid distances to spatial features}
\usage{
dist_to_categories(sf_obj, field, rasterlayer, n_jobs = 1)
}
\arguments{
\item{sf_obj}{sf object}

\item{field}{name of attribute that defines categories in the geomap to
calculate grid distances to}

\item{rasterlayer}{RasterLayer object to use as a template for the grid
distances}

\item{n_jobs}{number of processing cores, default = 1}
}
\value{
RasterStack of grid distances
}
\description{
This function splits a simple features object based on an attribute field
and rasterizes each category. Grid distances are then calculated to each
categorical raster separately. This function represents an alternative to
one-hot encoding or using rasters that represent categorical features.
However, instead of those features have distinct boundaries, the boundaries
here represent 'soft' margins defined by the proximity to the feature. This
avoids imprints of the polygons occurring in spatial predictions.
}
