% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distances.R
\name{dist_to_clusters}
\alias{dist_to_clusters}
\title{Calculate grid distances to x,y,z spatial clusters in point features}
\usage{
dist_to_clusters(sf_obj, field, rasterlayer, n = 10, iter = 100,
  n_jobs = 1)
}
\arguments{
\item{sf_obj}{sf data.frame containing POINT geometries}

\item{field}{character, name of field for z value}

\item{rasterlayer}{RasterLayer to use as a template}

\item{n}{integer, number of spatial clusters, default = 10}

\item{iter}{integer, maximum number of iterations for kmeans, default = 100}

\item{n_jobs}{integer, parallel processing, default = 1}
}
\value{
RasterStack of grid distances to x,y,z clusters
}
\description{
Calculates RasterLayers representing proximities to x,y,z spatial clusters
in simple features POINT geometries
}
