% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distances.R
\name{distance_to_features}
\alias{distance_to_features}
\title{Sample-based euclidean distance fields}
\usage{
distance_to_features(object, sf_obj)
}
\arguments{
\item{object}{RasterLayer to use as template}

\item{sf_obj}{Simple features object containing POINT geometries}
}
\value{
RasterStack of sample-based EDMs
}
\description{
Calculates sample-based euclidean distance fields, i.e. buffer distances
to each point in a simple features object
}
