% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterFunctions.R
\name{distanceFromPointIntervals}
\alias{distanceFromPointIntervals}
\title{Creates a RasterStack object containing euclidean distances to points}
\usage{
distanceFromPointIntervals(points, rasterlayer, field = NULL,
  n_classes = 10, method = "equal_intervals")
}
\arguments{
\item{points}{sf object}

\item{rasterlayer}{RasterLayer object to use as grid template}

\item{field}{character, name of field to group points into classes}

\item{n_classes}{numeric, number of classes}

\item{method}{character, either 'equal_intervals' or 'quantiles'
Used to create groups of points based on the field attribute}
}
\value{
RasterStack object of point distances
}
\description{
Creates a RasterStack object containing euclidean distances to points
}
