% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_visualization.R
\name{hillshade_pca}
\alias{hillshade_pca}
\title{Hillshade PCA function}
\usage{
hillshade_pca(x, azi = seq(0, 180, 22.5)[1:8], n = 5000)
}
\arguments{
\item{x}{RasterLayer object representing a DEM}

\item{azi}{numeric vector of azimuth of different shading directions}

\item{n}{numeric, number of pixels to sample for PCA. Default = 5000}
}
\value{
list, containing hillshade pca and the associated prcomp class
}
\description{
Hillshade PCA function
}
