% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{kernel_density2d}
\alias{kernel_density2d}
\title{2d kernel density estimation for raster data}
\usage{
kernel_density2d(data.points, y = NULL, xcells = NULL, ycells = NULL)
}
\arguments{
\item{data.points}{SpatialPointsDataFrame object}

\item{y}{RasterLayer object to use as a template for the output of the
kernel density estimator, optional.}

\item{xcells}{number of grid cells in x dimension of output raster}

\item{ycells}{number of grid cells in y dimension of output raster}
}
\value{
RasterLayer with KDE
}
\description{
2d kernel density estimation for raster data
}
