% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelDensity2D.R
\name{kernelDensity2D}
\alias{kernelDensity2D}
\title{2d kernel density estimation for raster data}
\usage{
kernelDensity2D(data.points, y = NULL, xcells = NULL, ycells = NULL)
}
\arguments{
\item{data.points}{SpatialPointsDataFrame object}

\item{y}{RasterLayer object to use as a template for the output of the
kernel density estimator, optional.}

\item{xcells}{number of grid cells in x dimension of output raster}

\item{ycells}{number of grid cells in y dimension of output raster}
}
\description{
2d kernel density estimation for raster data
}
