% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{merge_raster_factors}
\alias{merge_raster_factors}
\title{Merge factors containing a small number of observations}
\usage{
merge_raster_factors(x, y, n = 50)
}
\arguments{
\item{x}{RasterLayer object containing factors}

\item{y}{vector of extracted values to be used as training data}

\item{n}{integer, minimum number of observations required to retain a factor}
}
\value{
list
levels : data.frame with 3 columns containing raster values (ID),
the original levels that are associated with each value,
and  the collapsed levels
}
\description{
Merges factors containing a small number of observations
in training data extracted from a raster
}
