% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFilters.R
\name{morphological_filter}
\alias{morphological_filter}
\title{Morphological filter}
\usage{
morphological_filter(rasterlayer, size = 3, type = "closing",
  parallel = F)
}
\arguments{
\item{rasterlayer}{RasterLayer object}

\item{size}{Size of structuring element (in grid cells)}

\item{type}{Type of filter c('erosion', 'dilation', 'opening', 'closing)
Currently only closing is supported}

\item{parallel}{Logical, run focal function in parallel (default = FALSE)}
}
\value{
Rasterlayer after morphological filtering
}
\description{
Morphological filter
}
