% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{pred_fun_ranger}
\alias{pred_fun_ranger}
\title{Apply trained ranger model to a RasterStack or RasterBrick}
\usage{
pred_fun_ranger(model, data, factor_ldf = NULL, ...)
}
\arguments{
\item{model}{fitted ranger object}

\item{data}{passed from raster::predict function}

\item{factor_ldf}{optional, named list of dataframes containing
reclassification for factors in the raster data.  Each data.frame must
contain 3 columns, with column 1 containing raster values (ID),
column 2 containing the original levels that are associated with each raster 
value, and column 3 containing the collapsed levels}

\item{...}{additional named arguments to pass to the predict method}
}
\value{
RasterLayer object
}
\description{
Apply trained ranger model to a RasterStack or RasterBrick
}
