% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{ranger_tuned}
\alias{ranger_tuned}
\title{Hyperparameter tuning using the out-of-bag error of ranger models}
\usage{
ranger_tuned(object, formula, param_grid, data)
}
\arguments{
\item{object}{partial function of ranger}

\item{formula}{model formula}

\item{param_grid}{data.frame of hyperparameters, with names of
hyperparameters as columns and each trial as a row}

\item{data}{data.frame of training data}
}
\value{
list
best_model : model refitted using best hyperparameters
param_grid :  data.frame of param_grid with additional score column
}
\description{
Hyperparameter tuning using the out-of-bag error of ranger models
}
