% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertoVector.R
\name{rastertoVector}
\alias{rastertoVector}
\title{Raster to vector conversion with generalization}
\usage{
rastertoVector(rasterlayer, gapp = NULL, rmarea_km2, simplify_thres,
  smoothing_iter, cycles)
}
\arguments{
\item{rasterlayer}{RasterLayer object}

\item{gapp}{character, path to GRASS GIS}

\item{rmarea_km2}{Minimum size of polygons in km2}

\item{simplify_thres}{Neuuman reduction simplification threshold}

\item{smoothing_iter}{Number of chaiken smoothing iterations}

\item{cycles}{Number of reduction and smoothing iterations}
}
\value{
sf data.frame containing generalized polygons
}
\description{
Raster to vector conversion with generalization
}
