% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_distances.R
\name{rotated_grids}
\alias{rotated_grids}
\title{Produces RasterLayer objects filled with rotated coordinate values}
\usage{
rotated_grids(object, angles)
}
\arguments{
\item{object}{RasterLayer object}

\item{angles}{vector of angles to rotate coordinates by}
}
\value{
RasterStack object
}
\description{
Rotates the coordinate system around the x-axis by the specified angles
}
