% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterFunctions.R
\name{rotatedCoordinateGrids}
\alias{rotatedCoordinateGrids}
\title{Produces RasterLayer objects filled with rotated coordinate values}
\usage{
rotatedCoordinateGrids(object, n_angles)
}
\arguments{
\item{object}{RasterLayer object}

\item{n_angles}{numeric, number of angles to rotate coordinates by}
}
\value{
RasterStack object
}
\description{
Produces RasterLayer objects filled with rotated coordinate values
}
