% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterFunctions.R
\name{sampleEuclideanDistanceFields}
\alias{sampleEuclideanDistanceFields}
\title{Sample-based euclidean distance fields}
\usage{
sampleEuclideanDistanceFields(object, sf_obj, n_jobs = 1)
}
\arguments{
\item{object}{RasterLayer to use as template}

\item{sf_obj}{Simple features object containing POINT geometries}

\item{n_jobs}{numeric, optionally use parallel calculation over multiple cores}
}
\value{
RasterStack of sample-based EDMs
}
\description{
Calculates sample-based euclidean distance fields, i.e. buffer distances
to each point in a simple features object
}
