% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterFunctions.R
\name{xyCoordinateGrids}
\alias{xyCoordinateGrids}
\title{Creates RasterLayer objects filled by the x and y grid coordinates}
\usage{
xyCoordinateGrids(object)
}
\arguments{
\item{object}{RasterLayer object}
}
\value{
RasterStack object
}
\description{
Creates RasterLayer objects filled by the x and y grid coordinates
}
